clip-fast-api/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
uvicorn[standard]>=0.30.0,<0.31.0
pinecone-client>=2.2.4,<3.0.0
datasets>=2.18.0,<3.0.0
transformers>=4.38.2,<5.0.0
torch>=2.2.0,<3.0.0
Pillow>=10.2.0,<11.0.0
tqdm>=4.66.4,<5.0.0